+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
 ## vNext
+### Added
 - Setup for workflows containing a description that gets output
   
+### Changed
+- Breaking: Refactor Static StartWith methods from WorkflowBuilder to Workflow
+ 
 ## v1.0.1
 ### Added
 - Add multiple Assertions to Verificationstep